+++ src/components/FilterBar/index.js
@@ -54,7 +54,7 @@ const styles = {
         position: 'absolute',
         left: 0,
         right: 0,
-        bottom: 0,
+        bottom: -80,
         backgroundColor: 'rgba(0,0,0,0.4)',
         zIndex: 1
     }